requirements/prod.txt
flask==1.0.2
flask-restful==0.3.6
flask-jwt-extended==3.11.0
flask-sqlalchemy==2.3.2
flask-migrate==2.2.1
marshmallow==2.15.3
passlib==1.7.1
webargs==4.0.0
